api/requirements.txt
django
psycopg2-binary
Pillow
pylint
pylint-django
django-filter
djangorestframework
django-cors-headers
